refactor(project-structure): изменение структуры проекта
Jira issue: VEGA-194
