+++ src/pages/projects/ProjectsPage.test.tsx
@@ -1,10 +1,15 @@
 import React from 'react';
+import { BrowserRouter as Router } from 'react-router-dom';
 import * as tl from '@testing-library/react';
 
 import { ProjectsPage } from './ProjectsPage';
 
 function renderComponent(): tl.RenderResult {
-  return tl.render(<ProjectsPage />);
+  return tl.render(
+    <Router>
+      <ProjectsPage />
+    </Router>,
+  );
 }
 
 describe('ProjectsPage', () => {
