+++ src/ui/features/auth/AuthForm/GazpromLogo.tsx
