+++ src/ui/features/auth/AuthForm/cn-auth-form.ts
@@ -1,0 +1,3 @@
+import block from 'bem-cn';
+
+export const cnAuthForm = block('VegaAuthForm');
