+++ src/ui/features/auth/AuthForm/index.ts
@@ -1,0 +1,1 @@
+export { AuthForm } from './AuthForm';
